# src/main/resources/static/openapi-public.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Public Anime REST API",
    "description": "Public Anime API Docs",
    "version": "1.0.3"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/anime/search": {
      "get": {
        "tags": [
          "public anime"
        ],
        "summary": "Find anime by title",
        "description": "Search for anime using a title in English, Russian, or Japanese.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The title of the anime to search for."
          }
        ],
        "responses": {
          "200": {
            "description": "List of anime matching the search query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResult"
                }
              }
            }
          },
          "404": {
            "description": "Anime not found"
          }
        }
      }
    },
    "/user": {
      "post": {
        "tags": [
          "user data"
        ],
        "summary": "Add a new user",
        "operationId": "addUser",
        "requestBody": {
          "description": "New user information",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully created"
          },
          "409": {
            "description": "Conflict, e.g., email or username already taken"
          }
        }
      }
    },
    "/user/{username}": {
      "put": {
        "tags": [
          "user data"
        ],
        "summary": "Update user information",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Username of the user to be updated"
          }
        ],
        "requestBody": {
          "description": "Updated user data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User information successfully updated"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Conflict, e.g., email or username already taken"
          },
          "401": {
            "description": "Unauthorized - user does not have permission to update this user"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/genres": {
      "get": {
        "tags": [
          "genre list"
        ],
        "summary": "Get list of genres",
        "description": "Returns a list of all genres in the `RequestedGenreDTO` format.",
        "operationId": "getGenres",
        "responses": {
          "200": {
            "description": "Successful request. Returns a list of genres.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestedGenreDTO"
                }
              }
            }
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/anime": {
      "get": {
        "tags": [
          "public anime"
        ],
        "summary": "Get Anime List",
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "example": "finished",
            "required": false,
            "description": "Anime by status (upcoming, ongoing, finished)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "sortBy",
            "example": "score",
            "in": "query",
            "required": false,
            "description": "Sort by (score)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "genre",
            "example": "Action",
            "in": "query",
            "required": false,
            "description": "Required genres",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "type",
            "example": "TV",
            "in": "query",
            "required": false,
            "description": "Anime by type (TV, ONA, OVA, Special, Movie, Music)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "page",
            "example": 1,
            "in": "query",
            "required": false,
            "description": "Requested page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "example": 20,
            "in": "query",
            "required": false,
            "description": "Maximum elements per page",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of anime",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestedAnimeDTO"
                }
              }
            }
          },
          "400": {
            "description": "Incorrect attribute"
          },
          "404": {
            "description": "Anime not found"
          }
        }
      }
    },
    "/anime/{id}": {
      "get": {
        "tags": [
          "public anime"
        ],
        "summary": "Get Anime By Id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Anime ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Anime details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleAnimeDTO"
                }
              }
            }
          },
          "404": {
            "description": "Anime not found"
          }
        }
      }
    },
    "/user/{username}/myList/{status}": {
      "get": {
        "tags": [
          "user's anime"
        ],
        "summary": "Get User Anime By Status",
        "description": "Get a user's anime list by status (Watching, Rewatching, Onhold, Planned, All, Completed, Dropped).",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "description": "Username",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "path",
            "required": true,
            "description": "Anime status (Watching, Rewatching, Onhold, Planned, All, Completed, Dropped)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Requested page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum elements per page",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user anime",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestedAnimeDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User or anime not found"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/user/{username}/myList/add": {
      "post": {
        "tags": [
          "user's anime"
        ],
        "summary": "Add Anime To User List",
        "description": "Add anime to the user's list.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "description": "Username",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "Anime ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "description": "Status of the anime",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Anime added"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User or anime not found"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/user/{username}/myList/delete": {
      "delete": {
        "tags": [
          "user's anime"
        ],
        "summary": "Delete Anime From User List",
        "description": "Delete anime from the user's list.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "description": "Username",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "Anime ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Anime deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User or anime not found"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "SearchResult": {
        "type": "object",
        "properties": {
          "search_results": {
            "type": "integer",
            "description": "The number of search results."
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Anime"
            }
          }
        }
      },
      "UserDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Username"
          },
          "password": {
            "type": "string",
            "description": "Password"
          },
          "first_name": {
            "type": "string",
            "description": "First name"
          },
          "second_name": {
            "type": "string",
            "description": "Last name"
          },
          "patronymic": {
            "type": "string",
            "description": "Patronymic"
          },
          "email": {
            "type": "string",
            "description": "Email"
          }
        },
        "required": [
          "username",
          "password",
          "email"
        ]
      },
      "RequestedGenreDTO": {
        "type": "object",
        "properties": {
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          }
        },
        "required": [
          "genres"
        ]
      },
      "Genre": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Уникальный идентификатор жанра."
          },
          "name": {
            "type": "string",
            "description": "Название жанра."
          },
          "nameRus": {
            "type": "string",
            "description": "Название жанра на русском языке."
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "SingleAnimeDTO": {
        "type": "object",
        "description": "DTO for a single anime detail response.",
        "properties": {
          "id": {
            "type": "integer"
          },
          "image_url": {
            "type": "string"
          },
          "small_image_url": {
            "type": "string"
          },
          "large_image_url": {
            "type": "string"
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          },
          "studios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Studio"
            }
          },
          "trailer": {
            "$ref": "#/components/schemas/Trailer"
          },
          "title": {
            "type": "string"
          },
          "title_english": {
            "type": "string"
          },
          "title_japanese": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "episodes": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "airing": {
            "type": "boolean"
          },
          "aired_from": {
            "type": "string"
          },
          "aired_to": {
            "type": "string"
          },
          "aired_str": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "rating": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "float"
          },
          "synopsis": {
            "type": "string"
          },
          "season": {
            "type": "string"
          },
          "year": {
            "type": "string"
          }
        }
      },
      "RequestedAnimeDTO": {
        "type": "object",
        "description": "DTO for requested anime list response.",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Anime"
            }
          }
        }
      },
      "Pagination": {
        "type": "object",
        "description": "Pagination details.",
        "properties": {
          "last_page": {
            "type": "integer"
          },
          "has_next_page": {
            "type": "boolean"
          },
          "current_page": {
            "type": "integer"
          },
          "items": {
            "$ref": "#/components/schemas/Items"
          }
        }
      },
      "Items": {
        "type": "object",
        "description": "Items details for pagination.",
        "properties": {
          "count": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "per_page": {
            "type": "integer"
          }
        }
      },
      "Anime": {
        "type": "object",
        "description": "Anime details.",
        "properties": {
          "id": {
            "type": "integer"
          },
          "image_url": {
            "type": "string"
          },
          "small_image_url": {
            "type": "string"
          },
          "large_image_url": {
            "type": "string"
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          },
          "studios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Studio"
            }
          },
          "trailer": {
            "$ref": "#/components/schemas/Trailer"
          },
          "title": {
            "type": "string"
          },
          "title_english": {
            "type": "string"
          },
          "title_japanese": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "episodes": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "airing": {
            "type": "boolean"
          },
          "aired_from": {
            "type": "string"
          },
          "aired_to": {
            "type": "string"
          },
          "aired_str": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "rating": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "float"
          },
          "synopsis": {
            "type": "string"
          },
          "season": {
            "type": "string"
          },
          "year": {
            "type": "string"
          }
        }
      },
      "Trailer": {
        "type": "object",
        "description": "Trailer details.",
        "properties": {
          "youtube_id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "embed_url": {
            "type": "string"
          },
          "image_url": {
            "type": "string"
          },
          "small_image_url": {
            "type": "string"
          },
          "medium_image_url": {
            "type": "string"
          },
          "large_image_url": {
            "type": "string"
          },
          "max_image_url": {
            "type": "string"
          }
        }
      },
      "Studio": {
        "type": "object",
        "description": "Studio details.",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}